Add Initial DocumentServiceSecurityTest
